arquivos base agora estao separados, ainda sem modelagem
