+++ semana19/laBookProjeto19/src/model/post.ts
@@ -1,0 +1,14 @@
+export enum POST_TYPES {
+    NORMAL = "normal",
+    EVENT = "event"
+ }
+ 
+ export type post = {
+    id: string,
+    photo: string,
+    description: string,
+    type: POST_TYPES,
+    createdAt: Date,
+    authorId: string
+ }
+
