+++ semana19/laBookProjeto19/src/model/user.ts
@@ -1,0 +1,11 @@
+
+export type authenticationData = {
+    id: string
+}
+
+export type user = {
+    id: string,
+    name: string,
+    email: string,
+    password: string
+}
